Put minizinc related code in dedicated modules
- use "_mzn" in module name for easy recognition
- avoid warnings from minizinc when the binary is not installed
  and we are not using minizinc solvers
- NB: API breaks as module names for minzinc solvers changed
