+++ src/discrete_optimization/generic_rcpsp_tools/solvers/lns_cp_mzn/neighbor_builder.py
@@ -24,7 +24,7 @@ from discrete_optimization.generic_rcpsp_tools.solvers.lns_cp_mzn.neighbor_tools
     ParamsConstraintBuilder,
 )
 from discrete_optimization.generic_rcpsp_tools.typing import ANY_RCPSP
-from discrete_optimization.generic_tools.lns_cp import MznConstraintHandler
+from discrete_optimization.generic_tools.lns_cp_mzn import MznConstraintHandler
 from discrete_optimization.generic_tools.lns_tools import ConstraintHandlerMix
 from discrete_optimization.rcpsp_multiskill.problem import (
     MultiskillRcpspProblem,
