+++ src/discrete_optimization/rcpsp/solvers/cp_mzn.py
@@ -14,14 +14,16 @@ from minizinc import Instance, Model, Solver
 
 from discrete_optimization.generic_tools.cp_tools import (
     CpSolverName,
-    MinizincCpSolver,
     ParametersCp,
     SignEnum,
-    find_right_minizinc_solver_name,
 )
 from discrete_optimization.generic_tools.do_problem import (
     ParamsObjectiveFunction,
     build_aggreg_function_and_params_objective,
+)
+from discrete_optimization.generic_tools.mzn_tools import (
+    MinizincCpSolver,
+    find_right_minizinc_solver_name,
 )
 from discrete_optimization.rcpsp.problem import RcpspProblem
 from discrete_optimization.rcpsp.problem_preemptive import (
