+++ src/discrete_optimization/rcpsp_multiskill/solvers/cp_mzn.py
@@ -18,12 +18,14 @@ from discrete_optimization.generic_rcpsp_tools.graph_tools import (
 )
 from discrete_optimization.generic_tools.cp_tools import (
     CpSolverName,
-    MinizincCpSolver,
     ParametersCp,
     SignEnum,
-    find_right_minizinc_solver_name,
 )
 from discrete_optimization.generic_tools.do_problem import ParamsObjectiveFunction
+from discrete_optimization.generic_tools.mzn_tools import (
+    MinizincCpSolver,
+    find_right_minizinc_solver_name,
+)
 from discrete_optimization.rcpsp.problem_preemptive import PreemptiveRcpspSolution
 from discrete_optimization.rcpsp.solution import PartialSolution, RcpspSolution
 from discrete_optimization.rcpsp_multiskill.problem import (
