+++ src/discrete_optimization/rcpsp_multiskill/solvers/cp_mzn_mspsp_instlib.py
@@ -18,11 +18,13 @@ from discrete_optimization.generic_rcpsp_tools.graph_tools import (
 )
 from discrete_optimization.generic_tools.cp_tools import (
     CpSolverName,
-    MinizincCpSolver,
     SignEnum,
-    find_right_minizinc_solver_name,
 )
 from discrete_optimization.generic_tools.do_problem import ParamsObjectiveFunction
+from discrete_optimization.generic_tools.mzn_tools import (
+    MinizincCpSolver,
+    find_right_minizinc_solver_name,
+)
 from discrete_optimization.rcpsp_multiskill.problem import (
     MultiskillRcpspProblem,
     MultiskillRcpspSolution,
